Add the headers
There are two headers: one that defines the function attributes, and
one that undefines them -- the latter is expected to be `#include`d at
the end of a source file.

The attributes currently defined are: `INLINE` (the function's body is
expanded in replacement of a call), `NOTNULL()` (the specified
arguments must not be `NULL`), `PURE` (the function does not modify
external data) and `CONST` (like `PURE`, but the function does not even
access external data).