Define attribute DEFAULT
This attribute allows a function to be declared as a weak symbol, i.e.
designed to be overriden by an ulterior redeclaration.
This allows to provide default implementations to functions that user
code may or may not override (for example, UI event callback
functions).

The attribute's designed behavior is specific to GCC, however MSVC by
default behaves this way; for example, if a symbol is both defined in a
linked library and an object file, the symbol from the `.obj` is used
in preference.