Add attributes THISCALL and FASTCALL
The attributes modify the calling convention used for the function: the
first argument, if of integral type, is passed in the register `ECX`
and, for `FASTCALL`, the second in `EDX` -- again if of integral type.
Further arguments are passed on the stack, as usual.

The attributes are x86-specific, but are silently ignored on other
architectures (on x86_64, the convention is to pass the first four
arguments in the registers `RCX`, `RDX`, `R8` and `R9`).
Moreover they have no effect on variadic functions.

`THISCALL` is added to the definition of the OO-mimicking attribute
`MEMBER`.