Declare attribute DTOR
The purpose of the attribute is to provide a function to call right
before the variable falls out of scope and is to be discarded. It
allows to perform some clean up actions without having to call the
function explicitly.

The attribute is highly specific and only has a meaning under GNU C,
therefore it is only available if the test macro `HAVE_DTOR` is
defined. If this macro is defined and the compilation is not performed
in GNU mode, an error is raised.