Add variable attribute ALIGN(n)
The attribute permits to increase the size of the memory chunk used by
an element (variable, array or structure). It is usually used to add
padding bytes to members of a structure, or to enforce a larger size to
a variable for optimization purposes (processors are often more
efficient at reading portions of data of a certain size).

This attribute is defined in GNU C, Windows C and ISO C11.