Add I/O function attributes
The attributes are intended to provide advanced compile-time checks to
functions called with a `printf`-, `scanf`-, `strftime`- or
`strfmon`-like format string as an argument. The attributes allow the
syntax of the string to be checked as well as the types of the format
arguments.
The last argument specifies to the compiler that the function returns a
string consisting of a modified version of a format string taken as
argument. (This is useful to type check the function's arguments, as
well as removeing a GCC warning for a non-literal format string).
The attributes are GNU-C only, and are only defined if the macro
`HAVE_IOATTRS` is defined.